backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
pgvector==0.4.2
pydantic==2.10.0
python-dotenv==1.0.0

# ArXiv & PDF
arxiv==2.4.0
pypdf==6.6.0
requests==2.32.5

# NLP
nltk==3.9.2

# LangChain
langchain==1.2.3
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-community==0.4.1
langchain-text-splitters==1.1.0

# OpenAI
openai==2.15.0
